[ADD] NifTI-Loader can load HARDI Datasets
Note that for each HARDI dataset in NifTI-format, there must be an additional file
that contains the gradients. This file must end in .bvec and must otherwise have the same name
as the nifti file.

Example: <filename>.nii.gz  <filename>.bvec

The format of the gradient file is the same as output by the MRIConvert tool:
A text file with x-coordinates in the first line and y- and z- coordinates in the 2nd and last lines.