[ADD] This is a contribution from RRZK University of Cologne done by Martin Aumüller and Stefan Zellmann:
CUDA Implementation of the Zhang deterministic tract clustering. In order to use this clustering you should have a CUDA
capable GPU with 192 shader units.  This is only tested with GCC 4.3 and CUDA-SDK 2.3. Please note that CMake won't
complain if other versions are found. Do this at your own risk. In order to compile you will need to issue:

    `CC=/usr/bin/gcc-4.3 CXX=/usr/bin/g++-4.3 cmake ../../src`